+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Readonly text inputs with a default value are hidden.
+
 [4.52.2] - 2018-06-06
 ---------------------
 ##### Fixed